+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+* Prefetch links are now only added for transitive dependencies.
 <!-- Add new, unreleased changes here. -->
 
 ## [1.5.0] - 2017-05-23